+++ .npmignore
@@ -1,16 +1,21 @@
-/coverage
-/docs
+/demos
+/src
+/types
 /test
-/tools
 
 # misc
+/.dependabot
+/.github
 /.editorconfig
 /.env*
 /.eslintignore
-/.eslintrc.js
+/.eslintrc.yml
 /.git/
 /.gitignore
+/babel.config.js
 /yarn.lock
-/rollup.config.js
 /jest.config.js
+/rollup.config.js
+/tsconfig.*
 .gitkeep
+.idea
