+++ demos/playground/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "boxes",
+  "name": "playground",
   "version": "0.0.0",
   "description": "Boxes demo",
   "main": "app.js",
