+++ package.json
@@ -10,14 +10,12 @@
   "license": "MIT",
   "author": "Simon Ihmig <simon.ihmig@kaliber5.de>",
   "sideEffects": false,
-  "main": "dist/index.cjs.js",
-  "module": "dist/src/index.js",
-  "typings": "dist/types/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "main": "index.cjs.js",
+  "module": "index.js",
+  "typings": "index.d.ts",
   "scripts": {
-    "prebuild": "rimraf dist",
+    "clean": "rimraf index.js index.cjs.js index.d.ts *.map components systems -private",
+    "prebuild": "yarn clean",
     "build": "rollup -c rollup.config.js && tsc -b tsconfig.build.json",
     "docs": "typedoc --out docs --target es6 --theme minimal --mode file src",
     "lint": "yarn lint:ts && yarn lint:eslint",
@@ -25,7 +23,7 @@
     "lint:ts": "tsc --noEmit",
     "precommit": "lint-staged",
     "prepublishOnly": "yarn build",
-    "postpublish": "rimraf dist",
+    "postpublish": "yarn clean",
     "report-coverage": "cat ./coverage/lcov.info | coveralls",
     "start": "rollup -c rollup.config.js -w",
     "test": "jest",
@@ -87,6 +85,7 @@
     "prettier": "^2.1.0",
     "release-it": "^13.6.9",
     "release-it-lerna-changelog": "^2.3.0",
+    "rimraf": "^3.0.2",
     "rollup": "^2.26.5",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-sourcemaps": "^0.6.2",
