+++ src/components/position.ts
@@ -1,6 +1,6 @@
 import { Component, ComponentSchema } from 'ecsy';
 import { Vector3 } from '@babylonjs/core/Maths/math.vector';
-import { BabylonTypes } from '../ecsy-types';
+import { BabylonTypes } from '../-private/ecsy-types';
 
 export default class Position extends Component<Position> {
   value!: Vector3;
