+++ src/components/scale.ts
@@ -1,6 +1,6 @@
 import { Component, ComponentSchema } from 'ecsy';
 import { Vector3 } from '@babylonjs/core/Maths/math.vector';
-import { BabylonTypes } from '../ecsy-types';
+import { BabylonTypes } from '../-private/ecsy-types';
 
 export default class Scale extends Component<Scale> {
   value!: Vector3;
