+++ src/systems/action.ts
@@ -2,8 +2,8 @@
 
 import { Entity } from 'ecsy';
 import { Action, Mesh } from '../components';
-import SystemWithCore, { queries } from '../SystemWithCore';
-import assert from '../utils/assert';
+import SystemWithCore, { queries } from '../-private/SystemWithCore';
+import assert from '../-private/utils/assert';
 import { ActionManager } from '@babylonjs/core/Actions/actionManager';
 import { ExecuteCodeAction } from '@babylonjs/core/Actions/directActions';
 import { AbstractActionManager } from '@babylonjs/core/Actions/abstractActionManager';
