+++ src/systems/mesh.ts
@@ -1,7 +1,7 @@
 import { Entity } from 'ecsy';
 import { Mesh, TransformNode, Material } from '../components';
-import SystemWithCore, { queries } from '../SystemWithCore';
-import assert from '../utils/assert';
+import SystemWithCore, { queries } from '../-private/SystemWithCore';
+import assert from '../-private/utils/assert';
 import { AbstractMesh } from '@babylonjs/core/Meshes/abstractMesh';
 
 function detachFromScene(mesh: AbstractMesh): void {
